@@ -7,7 +7,9 @@
     "development": "nodemon"
   },
   "dependencies": {
-    "@hot-loader/react-dom": "16.13.0",
+    "@hot-loader/react-dom": "^17.0.1",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
@@ -21,18 +23,21 @@
     "lodash": "^4.17.21",
     "mongodb": "3.5.5",
     "mongoose": "^5.13.5",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-hot-loader": "4.12.20"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-hot-loader": "^4.13.0",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.14.8",
     "@babel/preset-env": "^7.14.8",
-    "@babel/preset-react": "7.9.4",
+    "@babel/preset-react": "^7.14.5",
     "@webpack-cli/generators": "^2.2.0",
     "babel-loader": "^8.2.2",
+    "file-loader": "^6.2.0",
     "nodemon": "2.0.2",
-    "webpack": "^5.47.1",
+    "webpack": "^5.50.0",
     "webpack-cli": "^4.7.2",
     "webpack-dev-middleware": "3.7.2",
     "webpack-dev-server": "^3.11.2",
