@@ -1,13 +1,18 @@
+import express from 'express'
+import path from 'path'
+import bodyParser from 'body-parser'
+import cookieParser from 'cookie-parser'
+import compress from 'compression'
+import cors from 'cors'
+import helmet from 'helmet'
 import userRoutes from './routes/user.routes'
-import authRouter from './routes/auth.routes'
+import authRoutes from './routes/auth.routes'
 
-const express = require ('express')
-const bodyParser = require('body-parser')
-const cookieParser = require('cookie-parser')
-const compress = require('compression')
-const cors = require('cors')
-const helmet = require('helmet')
-const app = express()
+//comment out before building for production
+import devBundle from './devBundle'
+
+const CURRENT_WORKING_DIR = process.cwd()
+const app = express() 
 
 app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: true }))
@@ -15,9 +20,13 @@ app.use(cookieParser())
 app.use(compress())
 app.use(helmet())
 app.use(cors())
+app.use('/dist', express.static(path.join(CURRENT_WORKING_DIR, 'dist')))
 
+//mount routes
 app.use('/', userRoutes)
-app.use('/', authRouter)
+app.use('/', authRoutes)
+
+devBundle.compile(app)
 
 //Auth error handling for express-jwt
 app.use((err, req, res, next) => {
